Add software-properties-gtk and software-properties-kde as provided packages.
mintsources conflicts with these packages, which leads to nasty dependency
handling behavior in apt. Since it provides exactly the same functionality,
and very few packages actually have these dependencies, we can specify that
we provide these packages and avoid any trouble or the need to manually build
copies of all sorts of things.