hacky workaround to support os.execl call
the change to os.execl from os.system messes up passing an undefined number of string arguments.  So I built in a "dummy" -? variable that does nothing, but is accepted and "eaten" by the parser.

blank arguments don't work - it passes a blank argument across, which messes up args[0].  Lists don't work - strings required by os.execl.  Trying to make a "tricky" comma separated string doesn't work - it is enclosed in double-quotes and treated as a single argument anyway - so the next parser doesn't accept the arguments.   A dummy  variable seemed better than a nested list of if's.  I'm trying to make it flexible so that it could handle additional arguments (more than --yes and --remove).