Fix crash in Select Mirror on no connection
Until now, we remove mirrors from the Mirror Selection dialog if they
aren't available. Sometimes this works fine but sometimes or if many
servers are unavailable e.g. due to no internet connection on the user
side, the list gets unsynchronized and crashes our script as removing
entries from a GtkListStore is not thread-safe.
Instead of making the whole thing thread-safe (with locks and
everything), this commit displays an "offline" label in the speed column
of such servers.

That way, the user is still able to select (seemingly) offline servers
but I consider this behavior quite appropriate and better than
presenting an empty list, which would be the consequence if the current
approach were thread-safe and the user started the tool without being
online.