Don't depend on gksu|kdesudo
Although these are needed, this dependency breaks the build for the KDE Edition.
Probably because it's pulled early, at a stage where the PPA isn't present
just yet.

gksu/kdesudo are both part of the default installation in their relevant
editions anyway.

The dependency on gksu is also set by gdebi, also part of the default install.