Also show packages already installed from a PPA (#105)
The PPA browser excluded packages if the version on the system was the same as in the PPA. This change also shows those packages, with a disabled (checked) checkbox. This way all the PPA's packages are shown in the PPA browser and only those not yet installed or upgradable can be marked for installation. Screenshot of the implementation: https://i.imgur.com/F1ENzXz.png