Add a cl argument to disable cache refresh, and use a return code to notify a calling program whether or not any sources changed.
When software-properties-gtk (mintsources) is called from synaptic,
its refresh dialog hangs, as synaptic is already holding a lock
on the cache.

Instead, support the '--no-update' flag that synaptic assumes it
has, and return 0 or 1 from the process to tell synaptic whether
or not it should prompt to update the cache itself.

Fixes https://github.com/linuxmint/lmde-5-cinnamon-beta/issues/29.