Add support for deb822 format
- To handle legacy/deb822 seamlessly, move to a patched version of
repolib
- Remove the edit buttons
- Remove deb-src support in 3rd party repos and PPAs

These removals are motivated by the fact that deb822 can have multiple
types per source. In other words, the concept of being enabled applies
to the whole source and both types..

as a result repolib doesn't handle deb-src well and ignores them
in legacy format when they're mixed in with deb lines.